requirements.txt
pandas==1.3.5
numpy==1.26.4
tensorboard==2.14.0
torch_geometric==2.3.1
torch-geometric-temporal==0.54.0
icecream==2.1.3